requirements.txt
django==4.2.2
djangorestframework==3.14.0
django-redis==5.3.0
celery==5.3.0